// src/store/branches/project/enums.ts
export enum ProjectActionType {
	ADD_PROJECT_DETAILS = 'ADD_PROJECT_DETAILS',
	FETCH_QUOTATION_DETAILS = 'FETCH_QUOTATION_DETAILS',
	STORE_QUOTATION_DETAILS = 'STORE_QUOTATION_DETAILS',

	FETCH_QUOTATION_LIST = 'FETCH_QUOTATION_LIST',
	STORE_QUOTATION_LIST = 'STORE_QUOTATION_LIST',

	GET_PRE_DESIGN_QUOTATION = 'GET_PRE_DESIGN_QUOTATION',
	STORE_PRE_DESIGN_QUOTATION = 'STORE_PRE_DESIGN_QUOTATION',
	ERROR_PRE_DESIGN_QUOTATION = 'ERROR_PRE_DESIGN_QUOTATION',

	GET_DEFAULT_QUOTATION = 'GET_DEFAULT_QUOTATION',
	STORE_DEFAULT_QUOTATION = 'STORE_DEFAULT_QUOTATION',
	ERROR_DEFAULT_QUOTATION = 'ERROR_DEFAULT_QUOTATION'
}